Connected all requests in main to the 'homes' app name. Made all mongo db connections linked to .env file. Entered 2 of 3 users on to the info in the .env file (are commented out)
Also reformated the main.py into a better readable top-down structure.

+++ DeviceSchema.py
@@ -4,5 +4,5 @@ class DeviceSchema(Schema):
     name = fields.String(required=True, validate=validate.Regexp("^[a-zA-Z0-9_+-]{3,}$"))
 
     #these are prone to change depneding on the actual format of the sensorIds
-    soundSensorId = fields.Int(required=True)
+    soundSensorId = fields.Float(required=True)
     motionSensorId = fields.Int(required=True)
